+++ config/test.exs
@@ -1,1 +1,3 @@
 use Mix.Config
+
+config :td_cache, permissions: []